pyproject.toml
[project]
name = "pdf-converter"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
    "fastapi>=0.115.0",
    "google-genai>=1.0.0",
    "jinja2>=3.1.0",
    "PyMuPDF>=1.24.0",
    "python-multipart>=0.0.12",
    "uvicorn>=0.30.0",
]
